On the StringHelper.cs, encoding an empty string returns an empty string instead of throwing an ArgumentNullException exception. This is useful when listing objects by PATH and the path is the root of the container.
With the StorageItemInformation Class, I added some obvious properties like SizeInBytes (as long instead of the ContentLength which was a String). To do this, I also changed the constructor of GetContainerInformation.

The Connection could give a list of objects as JSON. But there was no way to filter by path. I added a method to do this.

Signed-off-by: NZEYIMANA Emery Fabrice <nzem@nefsystems.com>